[BE] SearchStation API 구현, data.sql 편집, Station 필드 추가, Payment API 비즈니스 로직 수정 (#112)
* Fix: post, patch, delete, get Payment 비즈니스 로직 수정
(로그인검증, 시간검증, 총 가격 계산 로직 추가)

* Fix: Station confirmId 필드 추가

* Fix: cors이슈 수정 + webConfig 주석처리

* Docs: data.sql 편집

* Feat: 범위+시간 검색 Station API 구현

* Fix: Admin ROLE 부여

* Docs: Test data.sql

+++ backend/src/main/java/backend/domain/payment/service/PaymentService.java
@@ -18,19 +18,18 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.Duration;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.util.Optional;
 
 @Service @RequiredArgsConstructor @Transactional(readOnly = true)
 public class PaymentService {
 
     private final PaymentRepository paymentRepository;
-
     private final BatteryRepository batteryRepository;
-
     private final StationRepository stationRepository;
-
     private final MemberRepository memberRepository;
-
     private final ReservationRepository reservationRepository;
 
     @Transactional
@@ -44,25 +43,42 @@ public class PaymentService {
         Station station = stationRepository.findById(battery.getStation().getId())
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.STATION_NOT_FOUND));
 
-//        reservationRepository.find(battery.getReserve().getStartTime(), battery.getReserve().getEndTime())
-//        예약하려던 배터리가 가진 Reservation테이블의 값들 중 between시간이 현재 빌리려는 시간의 Btween시간과 겹치는지 비교하기
+        // 예약 시간 가능한지 검증 로직
+        for (int i = 0; i < battery.getReservations().size(); i++) {
+            if (((LocalDateTime.parse(battery.getReservations().get(i).getStartTime()).isAfter(LocalDateTime.parse(payment.getStartTime()))
+                    && LocalDateTime.parse(battery.getReservations().get(i).getStartTime()).isBefore(LocalDateTime.parse(payment.getEndTime()))))
+                    || (LocalDateTime.parse(battery.getReservations().get(i).getEndTime()).isBefore(LocalDateTime.parse(payment.getStartTime())))) {
+                throw new BusinessLogicException(ExceptionCode.CAN_NOT_RESERVE);
+            }
+        }
 
-        int totalPrice = battery.getPrice();  // *(endTime - startTime) 로직 추가되야함
+        // 총 금액 계산 로직
+        // 1. startTime과 endTime을 분단위로 환산 (절대 시간 계산?)
+        // String을 LocalDateTime으로 변환 -> 시간 비교 -> 분단위 환산  (endTime - startTime의 분단위 값)
+        LocalDateTime startTime = LocalDateTime.parse(payment.getStartTime());
+        LocalDateTime endTime = LocalDateTime.parse(payment.getEndTime());
+        Duration diff = Duration.between(startTime, endTime);
+        int diffMin = (int) diff.toMinutes();
+        // 2. 총 금액 = 기본 단위 가격 * 총 대여시간(min) / 10(min)
+        int totalPrice = battery.getPrice() * (diffMin / 10);
+
         payment.setMember(member);
         payment.setBattery(battery);
         payment.setStation(station);
         payment.setTotalPrice(totalPrice);
 
-//        결제가 불가능할 경우 예외처리 필요
-
         return paymentRepository.save(payment);
+
     }
 
 
     @Transactional
-    public Payment patchPayment (Payment payment) {
+    public Payment patchPayment (Payment payment, Long memberId) {
         Payment savedPayment = paymentRepository.findById(payment.getId())
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.PAY_NOT_FOUND));
+
+        if (savedPayment.getMember().getId() != memberId) throw new BusinessLogicException(ExceptionCode.NON_ACCESS_MODIFY);
+
         Optional.of(payment.getTotalPrice()).ifPresent(savedPayment::setTotalPrice);
         Optional.ofNullable(payment.getStatus()).ifPresent(savedPayment::setStatus);
         Optional.of(payment.getPayMethod()).ifPresent(savedPayment::setPayMethod);
@@ -85,18 +101,22 @@ public class PaymentService {
 
 
     @Transactional
-    public void deletePayment (Long paymentId) {
+    public void deletePayment (Long paymentId, Long memberId) {
         Payment savedPayment = paymentRepository.findById(paymentId)
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.PAY_NOT_FOUND));
+
+        if (savedPayment.getMember().getId() != memberId) throw new BusinessLogicException(ExceptionCode.NON_ACCESS_MODIFY);
 
         paymentRepository.delete(savedPayment);
     }
 
 
-    public Payment getPayment (Long paymentId) {
-
-        return paymentRepository.findById(paymentId)
+    public Payment getPayment (Long paymentId, Long memberId) {
+        Payment savedPayment = paymentRepository.findById(paymentId)
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.PAY_NOT_FOUND));
+        if (savedPayment.getMember().getId() != memberId) throw new BusinessLogicException(ExceptionCode.NON_ACCESS_MODIFY);
+
+        return savedPayment;
     }
